# api/docs/swagger.yaml
definitions:
  echo.HTTPError:
    properties:
      message: {}
    type: object
...
  github_com_USACE_instrumentation-api_api_internal_model.Measurement:
    properties:
      annotation:
        type: string
      error:
        type: string
      masked:
        type: boolean
      time:
        type: string
      validated:
        type: boolean
      value:
        type: number
    type: object
...
  github_com_USACE_instrumentation-api_api_internal_model.Timeseries:
    properties:
      id:
        type: string
      instrument:
        type: string
      instrument_id:
        type: string
      instrument_slug:
        type: string
      is_computed:
        type: boolean
      name:
        type: string
      parameter:
        type: string
      parameter_id:
        type: string
      project:
        type: string
      project_id:
        type: string
      project_slug:
        type: string
      slug:
        type: string
      unit:
        type: string
      unit_id:
        type: string
      values:
        items:
          $ref: '#/definitions/github_com_USACE_instrumentation-api_api_internal_model.Measurement'
        type: array
      variable:
        type: string
    type: object
  github_com_USACE_instrumentation-api_api_internal_model.TimeseriesCollectionItems:
    properties:
      items:
        items:
          $ref: '#/definitions/github_com_USACE_instrumentation-api_api_internal_model.Timeseries'
        type: array
    type: object
...
info:
  contact: {}
  description: Monitoring Instrumentation Data Acquisition Systems (MIDAS) Web API
  license:
    name: MIT
    url: https://github.com/USACE/instrumentation-api/blob/555ea51191ff1245fe5910a295862be7514aaec6/LICENSE.md
  title: MIDAS Web API
  version: "2.0"
paths:
...
  /projects/{project_id}/instruments/{instrument_id}/constants:
    get:
      parameters:
      - description: project uuid
        format: uuid
        in: path
        name: project_id
        type: string
      - description: instrument uuid
        format: uuid
        in: path
        name: instrument_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_USACE_instrumentation-api_api_internal_model.Timeseries'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: lists constants for a given instrument
      tags:
      - instrument-constant
    post:
      parameters:
      - description: project uuid
        format: uuid
        in: path
        name: project_id
        type: string
      - description: instrument uuid
        format: uuid
        in: path
        name: instrument_id
        required: true
        type: string
      - description: timeseries collection items payload
        in: body
        name: timeseries_collection_items
        required: true
        schema:
          $ref: '#/definitions/github_com_USACE_instrumentation-api_api_internal_model.TimeseriesCollectionItems'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_USACE_instrumentation-api_api_internal_model.Timeseries'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: creates instrument constants (i.e. timeseries)
      tags:
      - instrument-constant
  /projects/{project_id}/instruments/{instrument_id}/constants/{timeseries_id}:
    delete:
      parameters:
      - description: project uuid
        format: uuid
        in: path
        name: project_id
        type: string
      - description: instrument uuid
        format: uuid
        in: path
        name: instrument_id
        required: true
        type: string
      - description: timeseries uuid
        format: uuid
        in: path
        name: timeseries_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: removes a timeseries as an instrument constant
      tags:
      - instrument-constant
...
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"